ALL GHOSTS WORK WELL
fear mode, dead, random, NOT TIMERS FOR SCATTER AND CHASE MODE

+++ Pacman/Vector2f.h
@@ -69,7 +69,7 @@ public:
 
 	float Vector2f::Length() const
 	{
-		return sqrt(myX*myX + myY*myY);
+		return float(sqrt(myX*myX + myY*myY));
 	}
 
 	void Vector2f::Normalize()
